# agrotech_backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "AgroTech",
        "version": "1.0.0",
        "title": "AgroTech",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "yogeshwaran@doodlblue.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "host": "localhost:2122",
    "basePath": "/api/v1",
...
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/land/createLand": {
            "post": {
                "tags": [
                    "CreateLand"
                ],
                "summary": "response",
                "description": "CreateLand",
                "operationId": "CreateLand",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "response",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/createLand"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "example": {
                                "success": false,
                                "status": 400,
                                "message": "Bad Request"
                            }
                        }
                    },
                    "200": {
                        "description": "Successful Operation",
                        "schema": {
                            "example": {
                                "status": 200,
                                "message": "Land Created Successfully",
                                "data": {
                                    "_id": "622b385e69538366bd09495c",
                                    "surveyNo": "218-111",
                                    "districtName": "Vishakapatnam",
                                    "villageName": "Yalamanjili",
                                    "zoneName": "Yalamanjili",
                                    "areaUnits": "Cents",
                                    "nonCultivableArea": "0.0000",
                                    "cultivableArea": "6.1300",
                                    "natureOfEarth": "JirattiPatti",
                                    "discipline": "0.00",
                                    "landDescription": "Pallam",
                                    "reservoir": "Canal",
                                    "strategicArea": "6.1300",
                                    "accountNumber": "928",
                                    "graduateName": "Sethu",
                                    "fatherHusbandSpouseName": "Mani",
                                    "experienceArea": "1.0200",
                                    "natureOfExperience": "Purchase",
                                    "landType": "DryLand",
                                    "documentNumber": "ASDGH2768",
                                    "landRate": "109,000",
                                    "landId": "VIYAYA218-111",
                                    "createdAt": "2022-03-11T11:54:06.850Z",
                                    "updatedAt": "2022-03-11T11:54:06.850Z",
                                    "__v": 0
                                }
                            }
                        }
                    }
                }
            }
        },
        "/land/editLand": {
            "put": {
                "tags": [
                    "EditLand"
                ],
                "summary": "response",
                "description": "EditLand",
                "operationId": "EditLand",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "id",
                        "description": "response",
                        "required": true
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "response",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/editLand"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "example": {
                                "success": false,
                                "status": 400,
                                "message": "Bad Request"
                            }
                        }
                    },
                    "200": {
                        "description": "Successful Operation",
                        "schema": {
                            "example": {
                                "status": 200,
                                "message": "LandRecords Updated Successfully",
                                "data": {
                                    "_id": "6229e1a497d612a399c315ee",
                                    "surveyNo": "120",
                                    "districtName": "Vishakapatnam",
                                    "villageName": "Yalamanjili",
                                    "zoneName": "Yalamanjili",
                                    "areaUnits": "Cents",
                                    "nonCultivableArea": "0.0000",
                                    "cultivableArea": "6.1300",
                                    "natureOfEarth": "JirattiPatti",
                                    "discipline": "0.00",
                                    "landDescription": "Pallam",
                                    "reservoir": "Canal",
                                    "strategicArea": "6.1300",
                                    "accountNumber": "928",
                                    "graduateName": "Sethu",
                                    "experienceArea": "1.0200",
                                    "natureOfExperience": "Purchase",
                                    "landType": "DryLand",
                                    "documentNumber": "ASDGH2768",
                                    "landRate": "109,000",
                                    "createdAt": "2022-03-10T11:31:48.432Z",
                                    "updatedAt": "2022-03-10T11:31:48.432Z",
                                    "__v": 0
                                }
                            }
                        }
                    }
                }
            }
        },
        "/land/deleteLand": {
            "delete": {
                "tags": [
                    "DeleteLand"
                ],
                "summary": "response",
                "description": "DeleteLand",
                "operationId": "DeleteLand",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "id",
                        "description": "response",
                        "required": true
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "example": {
                                "success": false,
                                "status": 400,
                                "message": "Bad Request"
                            }
                        }
                    },
                    "200": {
                        "description": "Successful Operation",
                        "schema": {
                            "example": {
                                "status": 200,
                                "message": "Land Deleted Successfully",
                                "data": null
                            }
                        }
                    }
                }
            }
        },
        "/land/getLand": {
            "get": {
                "tags": [
                    "GetLand"
                ],
                "summary": "response",
                "description": "GetLand",
                "operationId": "GetLand",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "search",
                        "description": "response",
                        "required": false
                    },
                    {
                        "in": "query",
                        "name": "page",
                        "description": "response",
                        "required": false
                    },
                    {
                        "in": "query",
                        "name": "limit",
                        "description": "response",
                        "required": false
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "example": {
                                "success": false,
                                "status": 400,
                                "message": "Bad Request"
                            }
                        }
                    },
                    "200": {
                        "description": "Successful Operation",
                        "schema": {
                            "example": {
                                "status": 200,
                                "message": "Data Fetched Successfully",
                                "data": [
                                    {
                                        "_id": "6225b5a3e127b551aa89d09c",
                                        "surveyNo": "109",
                                        "districtName": "Vishakapatnam",
                                        "villageName": "Yalamanjili",
                                        "zoneName": "Yalamanjili",
                                        "areaUnits": "Cents",
                                        "nonCultivableArea": "0.0000",
                                        "cultivableArea": "6.1300",
                                        "natureOfEarth": "JirattiPatti",
                                        "discipline": "0.00",
                                        "landDescription": "Pallam",
                                        "reservoir": "Canal",
                                        "strategicArea": "6.1300",
                                        "accountNumber": "928",
                                        "graduateName": "Sethu",
                                        "spouseName": "Mani",
                                        "experienceArea": "1.0200",
                                        "natureOfExperience": "Purchase",
                                        "landType": "DryLand",
                                        "documentNumber": "ASDGH2768",
                                        "landRate": "109,000",
                                        "createdAt": "2022-03-07T07:34:59.470Z",
                                        "updatedAt": "2022-03-08T13:33:58.545Z",
                                        "__v": 0,
                                        "status": 0
                                    }
                                ]
                            }
                        }
                    }
                }
            }
        },
        "/land/getLandById": {
            "get": {
                "tags": [
                    "GetLandById"
                ],
                "summary": "response",
                "description": "GetLandById",
                "operationId": "GetLandById",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "id",
                        "description": "response",
                        "required": true
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "example": {
                                "success": false,
                                "status": 400,
                                "message": "Bad Request"
                            }
                        }
                    },
                    "200": {
                        "description": "Successful Operation",
                        "schema": {
                            "example": {
                                "status": 200,
                                "message": "Data Fetched Successfully",
                                "data": {
                                    "_id": "6225b5a3e127b551aa89d09c",
                                    "surveyNo": "109",
                                    "districtName": "Vishakapatnam",
                                    "villageName": "Yalamanjili",
                                    "zoneName": "Yalamanjili",
                                    "areaUnits": "Cents",
                                    "nonCultivableArea": "0.0000",
                                    "cultivableArea": "6.1300",
                                    "natureOfEarth": "JirattiPatti",
                                    "discipline": "0.00",
                                    "landDescription": "Pallam",
                                    "reservoir": "Canal",
                                    "strategicArea": "6.1300",
                                    "accountNumber": "928",
                                    "graduateName": "Sethu",
                                    "spouseName": "Mani",
                                    "experienceArea": "1.0200",
                                    "natureOfExperience": "Purchase",
                                    "landType": "DryLand",
                                    "documentNumber": "ASDGH2768",
                                    "landRate": "109,000",
                                    "createdAt": "2022-03-07T07:34:59.470Z",
                                    "updatedAt": "2022-03-08T13:33:58.545Z",
                                    "__v": 0,
                                    "status": 0
                                }
                            }
                        }
                    }
                }
            }
        },
        "/land/getCount": {
            "get": {
                "tags": [
                    "GetCount"
                ],
                "summary": "response",
                "description": "GetCount",
                "operationId": "GetCount",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "districtName",
                        "description": "response",
                        "required": true
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "example": {
                                "success": false,
                                "status": 400,
                                "message": "Bad Request"
                            }
                        }
                    },
                    "200": {
                        "description": "Successful Operation",
                        "schema": {
                            "example": {
                                "status": 200,
                                "message": "Data Fetched Successfully",
                                "data": {
                                    "totalZoneCount": 2,
                                    "totalVillageCount": 4,
                                    "zoneData": [
                                        {
                                            "zoneName": "Yalamanjili",
                                            "villageCount": 2
                                        },
                                        {
                                            "zoneName": "Yalamanjili1",
                                            "villageCount": 2
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
    "securityDefinitions": {
        "bearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "definitions": {
...
        "createLand": {
            "type": "object",
            "required": [
                "surveyNo",
                "districtName",
                "villageName",
                "zoneName",
                "areaUnits",
                "nonCultivableArea",
                "cultivableArea",
                "natureOfEarth",
                "discipline",
                "landDescription",
                "reservoir",
                "strategicArea",
                "accountNumber",
                "graduateName",
                "fatherHusbandSpouseName",
                "experienceArea",
                "natureOfExperience",
                "landType",
                "documentNumber",
                "landRate"
            ],
            "properties": {
                "surveyNo": {
                    "type": "string"
                },
                "districtName": {
                    "type": "string"
                },
                "villageName": {
                    "type": "string"
                },
                "zoneName": {
                    "type": "string"
                },
                "areaUnits": {
                    "type": "string"
                },
                "nonCultivableArea": {
                    "type": "string"
                },
                "cultivableArea": {
                    "type": "string"
                },
                "natureOfEarth": {
                    "type": "string"
                },
                "discipline": {
                    "type": "string"
                },
                "landDescription": {
                    "type": "string"
                },
                "reservoir": {
                    "type": "string"
                },
                "strategicArea": {
                    "type": "string"
                },
                "accountNumber": {
                    "type": "string"
                },
                "graduateName": {
                    "type": "string"
                },
                "fatherHusbandSpouseName": {
                    "type": "string"
                },
                "experienceArea": {
                    "type": "string"
                },
                "natureOfExperience": {
                    "type": "string"
                },
                "landType": {
                    "type": "string"
                },
                "documentNumber": {
                    "type": "string"
                },
                "landRate": {
                    "type": "string"
                }
            },
            "example": {
                "surveyNo": "218-111",
                "districtName": "Vishakapatnam",
                "villageName": "Yalamanjili",
                "zoneName": "Yalamanjili",
                "areaUnits": "Cents",
                "nonCultivableArea": "0.0000",
                "cultivableArea": "6.1300",
                "natureOfEarth": "JirattiPatti",
                "discipline": "0.00",
                "landDescription": "Pallam",
                "reservoir": "Canal",
                "strategicArea": "6.1300",
                "accountNumber": "928",
                "graduateName": "Sethu",
                "fatherHusbandSpouseName": "Mani",
                "experienceArea": "1.0200",
                "natureOfExperience": "Purchase",
                "landType": "DryLand",
                "documentNumber": "ASDGH2768",
                "landRate": "109,000"
            }
        },
...
    },
...
}